+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   New component in React : `Link`
 -   Added `onBlur` prop to `Select` component
+-   New component in React : `WebBookmark`
 
 ## [0.21.8][] - 2020-02-19
 
@@ -23,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Added `focusPoint` prop to `Thumbnail` component
+
 
 ## [0.21.6][] - 2020-02-13